requirements.txt
fastapi
uvicorn
python-dotenv
requests
chromadb
sentence-transformers
pandas
numpy
googlemaps
tqdm
pydantic
